[test] Split DaVinciClientTest blob transfer tests into separate classes for parallel execution
